ensure the number of velocity points is integer

+++ py/rvspecfit/fitter_ccf.py
@@ -65,7 +65,7 @@ def fit(specdata, config):
 
     maxvel = config.get('max_vel') or 1000
     # only search for CCF peaks from -maxvel to maxvel
-    nvelgrid = 2 * maxvel / (config.get('vel_step0') or 2)
+    nvelgrid = int(2 * maxvel / (config.get('vel_step0') or 2))
     # number of points on the ccf in the specified velocity range
 
     loglam = {}
